@@ -12,78 +12,78 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.1.1",
-    "@angular/cli": "^8.1.1",
-    "@angular/common": "^8.1.1",
-    "@angular/compiler": "^8.1.1",
-    "@angular/core": "^8.1.1",
-    "@angular/forms": "^8.1.1",
-    "@angular/platform-browser": "^8.1.1",
-    "@angular/platform-browser-dynamic": "^8.1.1",
-    "@angular/router": "^8.1.1",
-    "@ng-bootstrap/ng-bootstrap": "5.0.0",
-    "@ngtools/webpack": "^1.2.4",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
+    "@ng-bootstrap/ng-bootstrap": "5.1.4",
+    "@ngtools/webpack": "^8.3.19",
     "@swimlane/ngx-charts": "^12.0.1",
-    "@swimlane/ngx-datatable": "^15.0.2",
-    "acorn": "^6.2.0",
-    "angular-calendar": "0.27.13",
+    "@swimlane/ngx-datatable": "^16.0.2",
+    "acorn": "^7.1.0",
+    "angular-calendar": "0.28.0",
     "angular-file-uploader": "^5.0.2",
-    "angular-notifier": "^4.1.1",
+    "angular-notifier": "^6.0.1",
     "angular-plotly.js": "^1.4.2",
     "bootstrap": "4.3.1",
     "braces": "^3.0.2",
-    "chart.js": "2.8.0",
-    "chartist": "0.11.3",
-    "core-js": "^2.5.1",
-    "d3": "^5.9.7",
-    "date-fns": "^1.30.1",
+    "chart.js": "2.9.3",
+    "chartist": "0.11.4",
+    "core-js": "^2.5.0",
+    "d3": "^5.14.1",
+    "date-fns": "^2.7.0",
     "hoek": "6.1.3",
     "jquery": "3.4.1",
     "jquery-slimscroll": "1.3.8",
     "keycloak-angular": "^7.0.1",
-    "keycloak-js": "^7.0.1",
-    "lodash": "^4.17.14",
+    "keycloak-js": "^8.0.0",
+    "lodash": "^4.17.15",
     "ng-chartist": "4.1.0",
-    "ng-multiselect-dropdown": "^0.2.3",
+    "ng-multiselect-dropdown": "^0.2.6",
     "ng2-archwizard": "^2.1.0",
     "ng2-charts": "2.3.0",
-    "ng2-completer": "^3.0.2",
+    "ng2-completer": "^3.0.3",
     "ng2-dragula": "2.1.1",
     "ng2-smart-table": "1.5.0",
     "ng2-validation": "^4.2.0",
     "ngx-perfect-scrollbar": "8.0.0",
-    "ngx-quill": "7.1.0",
-    "ngx-toastr": "10.0.4",
+    "ngx-quill": "7.3.9",
+    "ngx-toastr": "11.2.1",
     "pace-js": "1.0.2",
     "plotly.js": "^1.51.1",
-    "popper.js": "^1.12.5",
-    "quill": "^1.3.6",
+    "popper.js": "^1.16.0",
+    "quill": "^1.3.7",
     "randomatic": "^3.1.1",
-    "rxjs": "^6.3.3",
-    "rxjs-compat": "^6.0.0-rc.0",
-    "zone.js": "~0.9.1"
+    "rxjs": "^6.5.3",
+    "rxjs-compat": "^6.5.3",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.801.1",
-    "@angular/compiler-cli": "^8.1.1",
-    "@angular/language-service": "^8.1.1",
+    "@angular-devkit/build-angular": "~0.803.19",
+    "@angular/cli": "^8.3.19",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
     "@types/chartist": "0.9.46",
-    "@types/jasmine": "~3.3.13",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/jquery": "^3.3.30",
-    "@types/node": "~12.6.6",
+    "@types/jasmine": "~3.4.6",
+    "@types/jasminewd2": "~2.0.8",
+    "@types/jquery": "^3.3.31",
+    "@types/node": "~12.12.8",
     "angular-6-datatable": "^0.8.0",
-    "codelyzer": "^5.0.1",
-    "jasmine-core": "~3.4.0",
+    "codelyzer": "^5.2.0",
+    "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.2.0",
-    "karma-chrome-launcher": "~3.0.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
+    "karma": "^4.4.1",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "^2.1.0",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.2",
     "protractor": "^5.4.1",
-    "ts-node": "~8.3.0",
-    "tslint": "~5.18.0",
-    "typescript": "~3.4.5"
+    "ts-node": "~8.5.2",
+    "tslint": "~5.20.1",
+    "typescript": "~3.7.2"
   }
 }
