@@ -8,7 +8,7 @@ import {
 import { PerfectScrollbarConfigInterface } from 'ngx-perfect-scrollbar';
 // import { KeycloakService } from 'keycloak-angular';
 import { Router } from '@angular/router';
-import { KeycloakService } from '../../core/auth/keycloak.service';
+// import { KeycloakService } from '../../core/auth/keycloak.service';
 import { GlobalService } from '../../modules/services/global.service';
 import { CustomLogger } from '../../modules/utils/CustomLogger';
 declare var $: any;
@@ -23,8 +23,8 @@ export class NavigationComponent implements AfterViewInit {
   CURRENT_USERNAME = "EDWARD";
     public config: PerfectScrollbarConfigInterface = {};
   constructor(private modalService: NgbModal, private router:Router, private Globals:GlobalService) {
-    console.log("11111....KeycloakService:::", KeycloakService.auth.authz);
-    console.log("2222.....KeycloakService.auth.authz.authServerUrl:::", KeycloakService.auth.authz.tokenParsed.preferred_username);
+    // console.log("11111....KeycloakService:::", KeycloakService.auth.authz);
+    // console.log("2222.....KeycloakService.auth.authz.authServerUrl:::", KeycloakService.auth.authz.tokenParsed.preferred_username);
     this.CURRENT_USERNAME = GlobalService.current_username;
   }
 
@@ -96,7 +96,7 @@ export class NavigationComponent implements AfterViewInit {
 
   onClickLogout(): void {
     console.log("logging out");
-    KeycloakService.logout();
+    // KeycloakService.logout();
     this.router.navigate(['home']);
   }
 }
