@@ -3,7 +3,7 @@ import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
 import { AppModule } from './app/app.module';
 import { environment } from './environments/environment';
-import { KeycloakService } from './app/core/auth/keycloak.service';
+// import { KeycloakService } from './app/core/auth/keycloak.service';
 
 if (environment.production) {
   enableProdMode();
@@ -14,9 +14,9 @@ platformBrowserDynamic()
   .catch(err => console.log(err));
 
 
-KeycloakService.init()
-  .then(() => {
-    const platform = platformBrowserDynamic();
-    platform.bootstrapModule(AppModule);
-  })
-  .catch(function (error) { console.log(JSON.stringify(error)); return; });
+// KeycloakService.init()
+//   .then(() => {
+//     const platform = platformBrowserDynamic();
+//     platform.bootstrapModule(AppModule);
+//   })
+//   .catch(function (error) { console.log(JSON.stringify(error)); return; });
